+++ src/form/form.jsx
@@ -129,14 +129,17 @@ export default class Form extends React.Component {
         device: PropTypes.oneOf(['phone', 'tablet', 'desktop']),
         /**
          * 是否开启内置的响应式布局 （使用ResponsiveGrid）
+         * @version 1.19
          */
         responsive: PropTypes.bool,
         /**
          * 是否开启预览态
+         * @version 1.19
          */
         isPreview: PropTypes.bool,
         /**
          * 是否使用 label 替换校验信息的 name 字段
+         * @version 1.20
          */
         useLabelForErrorMessage: PropTypes.bool,
         /**