+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Change Log 
 
+## [1.21.13](https://github.com/alibaba-fusion/next/compare/1.21.12...1.21.13) (2020-12-01)
+
+
+### Bug Fixes
+
+* **Checkbox:** fix disabled checkbox label color and cursor display, close [#2231](https://github.com/alibaba-fusion/next/issues/2231) ([#2471](https://github.com/alibaba-fusion/next/issues/2471)) ([1a1e69a](https://github.com/alibaba-fusion/next/commit/1a1e69a))
+* **Table:** virtual table doubled height, close [#2475](https://github.com/alibaba-fusion/next/issues/2475) ([5b96708](https://github.com/alibaba-fusion/next/commit/5b96708))
+
+
+
+
+## [1.21.12](https://github.com/alibaba-fusion/next/compare/1.21.11...1.21.12) (2020-11-27)
+
+
+### Bug Fixes
+
+* **Select:** google translate bug fix [#2424](https://github.com/alibaba-fusion/next/issues/2424) ([8656521](https://github.com/alibaba-fusion/next/commit/8656521))
+* **Step:** extend icon's animation duration, close [#2414](https://github.com/alibaba-fusion/next/issues/2414) ([df0f25f](https://github.com/alibaba-fusion/next/commit/df0f25f))
+* **Typescript:** update  Radio.Group  ([c51b0bb](https://github.com/alibaba-fusion/next/commit/c51b0bb))
+
+
+
+
 ## [1.21.11](https://github.com/alibaba-fusion/next/compare/1.21.10...1.21.11) (2020-11-20)